Split random navigation playback facade
